first iteration of adding stripe checkout (#33)

+++ api-boundary/src/parts/models.rs
@@ -17,8 +17,8 @@ pub struct Part {
     pub material: String,  // TODO: Extract to enum in api-boundary.
     pub tolerance: String, // TODO: Extract to enum in api-boudnary.
     pub quantity: u64,
-    pub unit_price: Option<f64>,
-    pub sub_total: Option<f64>,
+    pub unit_price: Option<u64>,
+    pub sub_total: Option<u64>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
@@ -46,7 +46,7 @@ impl Part {
     }
 }
 
-#[derive(Serialize_enum_str, Deserialize_enum_str, Clone, Debug)]
+#[derive(Serialize_enum_str, Deserialize_enum_str, Clone, Debug, PartialEq)]
 #[serde(rename_all = "snake_case")]
 pub enum PartStatus {
     AwaitingPricing,
